requirements.txt
appdirs==1.4.0
dj-static==0.0.6
Django==1.10.3
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-grappelli==2.9.1
django-stronghold==0.2.8
django-user-accounts==2.0.0
jsonfield==1.0.3
packaging==16.8
pinax-eventlog==1.1.2
pinax-theme-bootstrap==7.11.0
pinax-webanalytics==2.0.4
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
static3==0.7.0
